requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-filter==23.5
Pillow==10.2.0
reportlab==4.0.9
django-cors-headers==4.3.1
drf-spectacular==0.27.0